add auth guard to selectively block routes for auth
