+++ src/apps/admin/index.ts
@@ -1,1 +1,2 @@
+export * from './layout/Dashboard'
 export * from './routes'
