--- src/apps/auth/pages/MemberLogin.tsx
@@ -1,3 +1,0 @@
-export const MemberLogin = () => {
-  return <div>MemberLogin</div>
-}
