--- src/apps/auth/pages/MemberRegister.tsx
@@ -1,3 +1,0 @@
-export const MemberRegister = () => {
-  return <div>MemberRegister</div>
-}
