+++ src/apps/auth/pages/Register.scss
