+++ src/apps/auth/pages/index.ts
@@ -1,4 +1,2 @@
-export * from './AdminLogin'
-export * from './AdminRegister'
-export * from './MemberLogin'
-export * from './MemberRegister'
+export * from './Login'
+export * from './Register'
