+++ src/apps/auth/routes.tsx
@@ -1,6 +1,6 @@
 import type { RouteObject } from 'react-router-dom'
 import { Link } from 'react-router-dom'
-import { AdminLogin, AdminRegister, MemberLogin, MemberRegister } from './pages'
+import { Login, Register } from './pages'
 
 export const authRoutes: RouteObject[] = [
   {
@@ -10,56 +10,21 @@ export const authRoutes: RouteObject[] = [
         <strong>Routes:</strong>
         <ul>
           <li>
-            <Link to="admin">Admin</Link>
+            <Link to="login">Login</Link>
           </li>
           <li>
-            <Link to="members">Members</Link>
+            <Link to="register">Register</Link>
           </li>
         </ul>
       </div>
     ),
   },
   {
-    path: 'admin',
-    children: [
-      {
-        index: true,
-        element: (
-          <div>
-            Admin auth index.
-            <ul>
-              <li>
-                <Link to="login">Login</Link>
-              </li>
-              <li>
-                <Link to="register">Register</Link>
-              </li>
-            </ul>
-          </div>
-        ),
-      },
-      {
-        path: 'login',
-        element: <AdminLogin />,
-      },
-      {
-        path: 'register',
-        element: <AdminRegister />,
-      },
-    ],
+    path: 'login',
+    element: <Login />,
   },
   {
-    path: 'members',
-    element: <div>Members auth index. Login/register as Member/Guest?</div>,
-    children: [
-      {
-        path: 'login',
-        element: <MemberLogin />,
-      },
-      {
-        path: 'register',
-        element: <MemberRegister />,
-      },
-    ],
+    path: 'register',
+    element: <Register />,
   },
 ]
